fix: file Two pieces of the same data appear in the uploaded file
