+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alitajs/dform",
-  "version": "3.3.2",
+  "version": "3.3.3",
   "scripts": {
     "start": "dumi dev",
     "docs:build": "dumi build",
