+++ src/components/NomarFile/fileGroup.tsx
@@ -1,9 +1,7 @@
-import React, { FC, useState, useEffect, forwardRef, useImperativeHandle, ChangeEvent } from 'react';
+import React, { FC } from 'react';
 import { INomarFileProps, INomarFileItemProps } from './interface';
-import _ from 'lodash';
 import ClosePng from '../../assets/close.png';
 import './index.less';
-import { getRandom } from '../../utils';
 
 const prefixCls = 'alitajs-dform-file';
 
@@ -11,51 +9,17 @@ interface IFileGroupProps extends INomarFileProps {
   value?: INomarFileItemProps[];
 }
 
-type IFileGroupPropsRef = (props: IFileGroupProps & { ref: any }, ref: any) => React.ReactElement<any, any> | null;
-
-const FileGroup: IFileGroupPropsRef = forwardRef((props, ref) => {
+const FileGroup: FC<IFileGroupProps> = (props) => {
   const {
     value = [],
     onChange,
     onClick,
     alias = { id: 'id', title: 'title' },
-    upload,
   } = props;
-
-  // const [val, setVal] = useState(value);
-
-  const addFileChange = (e: ChangeEvent<HTMLInputElement> | any) => {
-    if (e.target.files) {
-      const fileList = Object.keys(e.target.files).map(
-        (item) => e.target.files[item],
-      );
-
-      const values = fileList.map(item => ({
-        title: item.name,
-        fileId: getRandom(),
-      }));
-      const allFiles = [...value, ...values];
-      if (upload) {
-        upload(fileList);
-      }
-      if (onChange) {
-        onChange(allFiles, values, 'add');
-      }
-    }
-  };
-
-  useImperativeHandle(
-    ref,
-    () => ({
-      addFileChange
-    }),
-
-  )
 
   const del = (index: number) => {
     const newData = Array.from(value);
     newData.splice(index, 1);
-    // setVal(newData);
     if (onChange) onChange(newData, value[index], 'delete');
   };
 
@@ -87,6 +51,6 @@ const FileGroup: IFileGroupPropsRef = forwardRef((props, ref) => {
       ))}
     </div>
   );
-});
+};
 
 export default FileGroup;
